Logical entites from JKG
Some of the server code was modified in JKG repo to comment out asserts in the icarus code if entity number > 1023
Instead of doing this, I added checks for the relevant icarus calls so that they will never be executed if entity number > 1023

target_scriptrunner is logical in JKG code, I made it non-logical.  In scriptrunner_run there are lots of references to Icarus functions, which would seem to fail if is a logical entity?

I made all entities that have a script_targetname field forced to non-logical.  Icarus has to see them and thats run in the engine, so they cant be past 1024?

+++ codemp/game/g_utils.c
@@ -222,26 +222,42 @@ NULL will be returned if the end of the list is reached.
 gentity_t *G_Find (gentity_t *from, int fieldofs, const char *match)
 {
 	char	*s;
+	int idx;
 
 	if (!from)
 		from = g_entities;
 	else
 		from++;
 
-	for ( ; from < &g_entities[level.num_entities] ; from++)
+	idx = from - g_entities;
+	if (idx >= MAX_GENTITIES)
+		goto dological;
+
+	for (; from < &g_entities[level.num_entities]; from++)
 	{
 		if (!from->inuse)
 			continue;
-		s = *(char **) ((byte *)from + fieldofs);
+		s = *(char **)((byte *)from + fieldofs);
 		if (!s)
 			continue;
-		if (!Q_stricmp (s, match))
+		if (!Q_stricmp(s, match))
+			return from;
+	}
+	from = &g_logicalents[0]; // 1st logical entity
+dological:
+	for (; from < &g_logicalents[level.num_logicalents]; from++)
+	{
+		if (!from->inuse)
+			continue;
+		s = *(char **)((byte *)from + fieldofs);
+		if (!s)
+			continue;
+		if (!Q_stricmp(s, match))
 			return from;
 	}
 
 	return NULL;
 }
-
 
 
 /*
@@ -683,10 +699,17 @@ void G_InitGentity( gentity_t *e ) {
 	e->inuse = qtrue;
 	e->classname = "noclass";
 	e->s.number = e - g_entities;
+	if (e->s.number < 1023) {
+		e->isLogical = qfalse;
+	}
+	else {
+		e->isLogical = qtrue;
+	}
 	e->r.ownerNum = ENTITYNUM_NONE;
 	e->s.modelGhoul2 = 0; //assume not
 
-	trap->ICARUS_FreeEnt( (sharedEntity_t *)e );	//ICARUS information must be added after this point
+	if (!e->isLogical)
+		trap->ICARUS_FreeEnt( (sharedEntity_t *)e );	//ICARUS information must be added after this point
 }
 
 //give us some decent info on all the active ents -rww
@@ -840,6 +863,54 @@ gentity_t *G_Spawn( qboolean essential ) {
 	return e;
 }
 
+// G_SpawnLogical: Creates a logical entity (ent nums 1024 to 4097)
+gentity_t *G_SpawnLogical(void) {
+	int			i, force;
+	gentity_t	*e;
+
+	e = NULL;	// shut up warning
+	i = 0;		// shut up warning
+	for (force = 0; force < 2; force++) {
+		// if we go through all entities and can't find one to free,
+		// override the normal minimum times before use
+		e = &g_entities[MAX_GENTITIES];
+		for (i = MAX_GENTITIES; i< (MAX_GENTITIES + level.num_logicalents); i++, e++) {
+			if (e->inuse) {
+				continue;
+			}
+
+			// the first couple seconds of server time can involve a lot of
+			// freeing and allocating, so relax the replacement policy
+			if (!force && e->freetime > level.startTime + 2000 && level.time - e->freetime < 1000)
+			{
+				continue;
+			}
+
+			// reuse this slot
+			G_InitGentity(e);
+			return e;
+		}
+		if (i != MAX_ENTITIESTOTAL) {
+			break;
+		}
+	}
+	if (i == MAX_ENTITIESTOTAL - 1) {
+		/*
+		for (i = 0; i < MAX_GENTITIES; i++) {
+		trap->Print("%4i: %s\n", i, g_entities[i].classname);
+		}
+		*/
+		//G_SpewEntList();
+		trap->Error(ERR_DROP, "G_SpawnLogical: no free entities");
+	}
+
+	// open up a new slot
+	level.num_logicalents++;
+
+	G_InitGentity(e);
+	return e;
+}
+
 /*
 =================
 G_EntitiesFree
@@ -919,6 +990,7 @@ Marks the entity as free
 */
 void G_FreeEntity( gentity_t *ed ) {
 	//gentity_t *te;
+	int i, entnum;
 
 	if (ed->isSaberEntity)
 	{
@@ -928,9 +1000,11 @@ void G_FreeEntity( gentity_t *ed ) {
 		return;
 	}
 
-	trap->UnlinkEntity ((sharedEntity_t *)ed);		// unlink from world
-
-	trap->ICARUS_FreeEnt( (sharedEntity_t *)ed );	//ICARUS information must be added after this point
+	if (!ed->isLogical)
+	{
+		trap->UnlinkEntity((sharedEntity_t *)ed);		// unlink from world
+		trap->ICARUS_FreeEnt((sharedEntity_t *)ed);	//ICARUS information must be added after this point
+	}
 
 	if ( ed->neverFree ) {
 		return;
@@ -1028,6 +1102,32 @@ void G_FreeEntity( gentity_t *ed ) {
 	ed->classname = "freed";
 	ed->freetime = level.time;
 	ed->inuse = qfalse;
+
+	//Logical Entities - (JKG)
+	// Ok, lets see if we can lower level.num_entities.
+	// If this entity was the last allocated slot, run back through g_entities and get the last used slots.
+	entnum = ed - g_entities;
+	if (!ed->isLogical) {
+		if (entnum == level.num_entities - 1) {
+			// Last slot, roll back
+			for (i = entnum; i >= MAX_CLIENTS; i--) {
+				if (g_entities[i].inuse)
+					break;
+			}
+			level.num_entities = i + 1;
+		}
+	}
+	else {
+		if (entnum == MAX_GENTITIES + level.num_logicalents - 1) {
+			// Last slot, roll back
+			for (i = entnum; i >= MAX_GENTITIES; i--) {
+				if (g_entities[i].inuse)
+					break;
+			}
+			level.num_logicalents = i + 1 - MAX_GENTITIES;
+		}
+
+	}
 }
 
 /*
